Rename api v20201232preview -> v20201031preview
