+++ pkg/api/v20201031preview/generate.go
@@ -1,0 +1,6 @@
+//go:generate go run ../../../hack/swagger ../../../swagger/redhatopenshift/resource-manager/Microsoft.RedHatOpenShift/stable/2020-10-31-preview
+
+package v20201031preview
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the Apache License 2.0.
