+++ pkg/api/v20201031preview/openshiftcluster_marshal.go
@@ -1,4 +1,4 @@
-package v20201231preview
+package v20201031preview
 
 // Copyright (c) Microsoft Corporation.
 // Licensed under the Apache License 2.0.
