--- pkg/api/v20201231preview/generate.go
@@ -1,6 +1,0 @@
-//go:generate go run ../../../hack/swagger ../../../swagger/redhatopenshift/resource-manager/Microsoft.RedHatOpenShift/stable/2020-12-31-preview
-
-package v20201231preview
-
-// Copyright (c) Microsoft Corporation.
-// Licensed under the Apache License 2.0.
